networkをモジュールに分割
